+++ quasar.conf.js
@@ -45,7 +45,10 @@ module.exports = function (ctx) {
     },
     devServer: {
       // https: true,
+      host: '0.0.0.0',
+      disableHostCheck: true,
       port: 3000,
+      public: '10.0.1.61:3000',
       open: false // opens browser window automatically
     },
     // framework: 'all' --- includes everything; for dev only!
@@ -90,7 +93,9 @@ module.exports = function (ctx) {
         'QModal',
         'QPopover',
         'QVideo',
-        'QResizeObservable'
+        'QResizeObservable',
+        'QFab',
+        'QFabAction'
       ],
       directives: [
         'Ripple'